CustomJsonParser moved to common; group added to search page
