+++ backend/src/main/java/ticktrack/controllers/SearchController.java
@@ -10,7 +10,7 @@ import ticktrack.proto.Msg;
 
 import java.util.List;
 
-import static ticktrack.util.CustomJsonParser.*;
+import static common.helpers.CustomJsonParser.*;
 import static ticktrack.util.ResponseHandler.*;
 
 @Controller
