+++ backend/src/main/java/ticktrack/controllers/UserGroupController.java
@@ -5,12 +5,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import ticktrack.entities.UserGroup;
 import ticktrack.managers.UserGroupManager;
 import ticktrack.proto.Msg;
 
-import static ticktrack.util.CustomJsonParser.jsonToProtobuf;
-import static ticktrack.util.CustomJsonParser.protobufToJson;
+import static common.helpers.CustomJsonParser.jsonToProtobuf;
+import static common.helpers.CustomJsonParser.protobufToJson;
 import static ticktrack.util.ResponseHandler.buildFailureResponse;
 import static ticktrack.util.ResponseHandler.wrapCommonResponseIntoMsg;
