+++ frontend/src/main/java/ticktrack/frontend/controller/SearchController.java
@@ -3,6 +3,7 @@ package ticktrack.frontend.controller;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.protobuf.util.JsonFormat;
+import common.helpers.CustomJsonParser;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
@@ -16,6 +17,8 @@ import ticktrack.proto.Msg;
 
 import java.io.IOException;
 import java.util.List;
+
+import static common.helpers.CustomJsonParser.*;
 
 @Controller
 public class SearchController {
@@ -32,22 +35,45 @@ public class SearchController {
         Request categoriesRequest = new Request.Builder()
                 .url("http://localhost:9001/backend/v1/categories/getAllActive")
                 .build();
+        Request groupsRequest = new Request.Builder()
+                .url("http://localhost:9001/backend/v1/userGroups/getAll")
+                .build();
 
-        try (Response response = httpClient.newCall(categoriesRequest).execute()) {
-            if(response.code()==200) {
-                Msg.Builder builder = Msg.newBuilder();
-                JsonFormat.parser().merge(response.body().string(), builder);
-                Msg result = builder.build();
+        try (Response categoryResponse = httpClient.newCall(categoriesRequest).execute();
+        Response groupResponse = httpClient.newCall(groupsRequest).execute()) {
+            if(categoryResponse.code()==200) {
+                Msg result = jsonToProtobuf(categoryResponse.body().string());
 
                 model.put("categoryList", result.getCategoryOperation().getCategoryOpGetAllResponse().getCategoryNameList());
             } else {
-                logger.warn("Error received from backend, unable to get categories list: {}", response.message());
+                logger.warn("Error received from backend, unable to get categories list: {}", categoryResponse.message());
+            }
+
+            if(groupResponse.code() == 200) {
+                Msg result = jsonToProtobuf(groupResponse.body().string());
+
+                model.put("groupList", result.getUserGroupOperation().getUserGroupOpGetAllResponse().getGroupNameList());
+            } else {
+                logger.warn("Error received from backend, unable to get group list: {}", groupResponse.message());
             }
         } catch (IOException e) {
             logger.error("Internal error, unable to get categories list", e);
         }
 
         return "searchTicket";
+    }
+
+    @RequestMapping(value = "searchTickets", method = RequestMethod.GET)
+    @ResponseBody
+    List<Msg.TicketInfo> searchTickets(@RequestBody String jsonRequest) {
+
+        System.out.println("JSON RESULT:\n" + jsonRequest);
+
+        Request usersRequest = new Request.Builder()
+                .url("http://localhost:9001/backend/v1/searchUsers/")
+                .build();
+
+        return null;
     }
 
     @RequestMapping(value = "searchUsers", method = RequestMethod.GET)
